requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.9
customtkinter==5.2.2
pillow==10.2.0
numpy==1.26.3
protobuf==4.25.2
attrs==23.2.0
absl-py==2.1.0
flatbuffers==23.5.26
matplotlib==3.8.2
packaging==23.2
sounddevice==0.4.6
darkdetect==0.8.0 